refactor: move resize handle into shared module
