+++ packages/editor/src/shared/resize-handle.tsx
